### 2022-04-01 #4285 @yannikschaelte
Hi @kyleniemeyer @osorensen ! Sure, I will address the missing DOIs. Should I then simply update the paper.bib file, or should additional actions be performed?

Regarding reviewers, I think especially persons working on `statistical inference`, `parameter inference` and(/or) `systems/computational biology` (nhejazi, krother, khinsen, cthoyt, dirmeier, blakeaw, GregoryAshton, mattpitkin, marouenbg, dsurujon) or specifically `approximate Bayesian computation / likelihood-free inference` (janfb, matt-graham) as well as `Python` software development experience could be a good fit.
