### 2022-04-01 #4285 @osorensen
Thanks @yannikschaelte. Once you've updated `paper.bib`, you can confirm that it works by writing the comments `@editorialbot generate pdf` and `@editorialbot check references` in this issue.
