+++ doc/joss_0_12/paper.md
@@ -63,7 +63,7 @@ ABC is often combined with a sequential Monte-Carlo (ABC-SMC) scheme using impor
 ![**Basic ABC algorithm.** Parameters $\theta\sim\pi(\theta)$ are sampled from the prior or a proposal distribution, and passed to a black-box model generating potentially stochastic simulated data according to the likelihood $y\sim\pi(y|\theta)$. These are optionally passed through a summary statistics function giving a low-dimensional representation $s(y)$. Summary statistics of simulated and observed data are compared via a distance metric $d$, and the underlying parameters accepted if the distance is below an acceptance threshold $\varepsilon$.\label{fig:concept}](concept.pdf){ width=99% }
 
 While conceptually simple and widely applicable, ABC is computationally expensive, and its practical performance relies on a number of factors.
-There exist several software packages implementing different algorithms in different languages, each with their own strengths, including notably in Python ABCpy [@DuttaSch2017] and ELFI [@KangasraasioLin2016], in Julia GpABC [@TankhilevichIsh2020], and in R EasyABC [@JabotFau2013].
+There exist several software packages implementing different algorithms in different languages, each with their own strengths, including notably in Python ABCpy [@DuttaSch2017] and ELFI [@LintusaariVuo2018], in Julia GpABC [@TankhilevichIsh2020], and in R EasyABC [@JabotFau2013].
 pyABC implements at its core an ABC-SMC scheme based on @ToniStu2010 and facilicates robust and efficient inference for a broad spectrum of applications via robust methods and self-tuned choices of hyperparameters, reducing the need for manual tuning.
 An article on core features of pyABC was previously published [@KlingerRic2018], discussing in particular adaptive transition kernels [@FilippiBar2013], population sizes [@KlingerHas2017], and wall-time efficient parallelization via dynamic scheduling. pyABC is in use in a variety of fields, e.g. to model virus transmission on cellular [@ImleKum2019] and population level [@KerrStu2021], neuron circuits [@BittnerPal2021], cancer [@ColomHer2021], gene expression [@CoulierHel2021], axolotl regeneration [@CostaOts2021], universe expansion [@BernardoSai2021], cardiac electrophysiology [@CantwellYum2019], and bee colonies [@MinucciCur2021].
 
