+++ doc/joss_0_12/paper.bib
@@ -7,8 +7,6 @@
 
 @article{Kitano2002,
 	author = {Kitano, H.},
-	date-added = {2014-01-10 09:40:47 +0000},
-	date-modified = {2016-08-30 06:43:26 +0000},
 	journal = {Science},
 	keywords = {modeling; systems biology},
 	month = {Mar.},
@@ -18,12 +16,11 @@
 	title = {Systems Biology: A Brief Overview},
 	volume = {295},
 	year = {2002},
+	doi = {10.1126/science.1069492},
 }
 
 @book{Tarantola2005,
 	author = {Tarantola, A.},
-	date-added = {2012-04-17 10:57:01 +0200},
-	date-modified = {2013-03-11 13:44:45 +0100},
 	keywords = {parameter estimation, inverse problem, Bayesian parameter estimation, scaling, transformation, Monte-Carlo method, least squares},
 	publisher = {{SIAM}},
 	rating = {5},
@@ -40,6 +37,7 @@
 	title = {Inferring coalescence times from {DNA} sequence data},
 	volume = {145},
 	year = {1997},
+	doi = {10.1093/genetics/145.2.505},
 }
 
 @Article{Wilkinson2009,
@@ -48,14 +46,12 @@
   number        = {2},
   pages         = {122--133},
   volume        = {10},
-  bdsk-url-1    = {http://dx.doi.org/10.1038/nrg2509},
-  date-added    = {2011-06-18 20:29:26 +0200},
-  date-modified = {2013-04-24 18:50:43 +0200},
   journal       = {Nat. Rev. Genet.},
   keywords      = {SDE, stochastic modeling, parameter estimation, review},
   month         = {Feb.},
   rating        = {5},
   year          = {2009},
+  doi           = {10.1038/nrg2509},
 }
 
 @Article{JagiellaRic2017,
@@ -64,9 +60,7 @@
   number        = {2},
   pages         = {194--206},
   volume        = {4},
-  bdsk-url-1    = {http://dx.doi.org/10.1016/j.cels.2016.12.002},
-  date-added    = {2017-03-14 11:13:37 +0000},
-  date-modified = {2017-12-07 11:19:05 +0000},
+  doi           = {10.1016/j.cels.2016.12.002},
   journal       = {Cell Systems},
   keywords      = {parameter estimation, ABC, multi-scale modeling, agent-based models, basic reading},
   month         = {02},
@@ -82,6 +76,7 @@
   journal   = {The Annals of Statistics},
   publisher = {Institute of Mathematical Statistics},
   year      = {2009},
+  doi       = {10.1214/07-AOS574},
 }
 
 @article{GourierouxMon1993,
@@ -94,6 +89,7 @@
 	title = {Indirect inference},
 	volume = {8},
 	year = {1993},
+	doi = {10.1002/jae.3950080507},
 }
 
 @article{PriceDro2018,
@@ -106,6 +102,7 @@
 	title = {Bayesian synthetic likelihood},
 	volume = {27},
 	year = {2018},
+	doi = {10.1080/10618600.2017.1302882},
 }
 
 @Article{PritchardSei1999,
@@ -117,12 +114,11 @@
   journal   = {Molecular biology and evolution},
   publisher = {Oxford University Press},
   year      = {1999},
+  doi       = {10.1093/oxfordjournals.molbev.a026091},
 }
 
 @article{BeaumontZha2002,
 	author = {Beaumont, M. A. and Zhang, W. and Balding, D. J.},
-	date-added = {2012-01-20 10:46:56 +0100},
-	date-modified = {2015-05-16 16:11:25 +0000},
 	journal = {Genetics},
[…]
