### 2022-04-04 14ec4d5
Add example documenting custom priors
