### 2022-04-05 d96f1b4
Add example documenting custom priors (#559)
