### 2022-04-05 ae13bf9
update installation guide (#561)
* update installation guide

* update changelog

* add link
