+++ CHANGELOG.rst
@@ -8,6 +8,12 @@ Release Notes
 ...........
 
 
+0.12.3 (2022-04-05)
+-------------------
+
+* Document custom priors (#559)
+
+
 0.12.2 (2022-03-25)
 -------------------
 
