### 2022-04-07 #4304 @editorialbot
[REVIEW]: pyABC: Efficient and robust easy-to-use approximate Bayesian computation
**Submitting author:** @yannikschaelte (<a href="http://orcid.org/0000-0003-1293-820X">Yannik Schälte</a>)
**Repository:** https://github.com/icb-dcm/pyabc
**Branch with paper.md** (empty if default branch): joss_0_12
**Version:** 0.12.5
