### 2022-05-03 ICB-DCM/pyABC#565 @blakeaw
JOSS Review - Error when running automated tests
Hi,

This issue is part of the JOSS Review at https://github.com/openjournals/joss-reviews/issues/4304

I ran into an error when I tried to run the base automated tests with `tox -e base` as described in the docs at https://pyabc.readthedocs.io/en/latest/contribute.html#unit-tests

After then running `pytest base` I saw I was missing r2py and pyarrow, which are dependencies for the R interface `pip install pyabc[R]`.

A couple of comments:
1. Maybe pytest should be part of the dev requirements installed with `pip install -r requirements-dev.txt` since it doesn't seem to be installed otherwise. Or it could be noted more clearly in docs that you need to install pytest to locally run automated tests (even when doing so via tox).
2.  Since the R interface is an optional part of the package it might be helpful to note in the Contribute section of the docs that it has coverage in the base tests (`tox -e base` or `pytest base`) and needs to be installed for the relevant tests to pass (along with an R distribution).
