### 2022-05-03 ICB-DCM/pyABC#566 @blakeaw
JOSS Review - comment on Software paper: State of the field
Hi,

This issue is part of the JOSS Review at https://github.com/openjournals/joss-reviews/issues/4304

This is a minor comment for the Software paper itself in regard to the `State of the field` :

In the Statement of Need section, you note 'There exist several software packages implementing different algorithms in different languages, each with their own strengths,'

I think it might be helpful if you could expand on this statement a bit to clarify more what distinguishes pyABC from the packages you note (ABCpy, ELFI, etc.), particularly, for the competing Python implementations.

For example, maybe you could state what the main strengths (or weaknesses, if any) of those competing packages are, so it's more clear what gaps pyABC fills and how it compares to those other packages. Do the other packages not implement ABC-SMC? Are the new pyABC features you are introducing in the Software paper (or previous core features, or some subset of features) only available in pyABC?
