### 2022-05-03 ICB-DCM/pyABC#565 @yannikschaelte
Thanks for the issue! Indeed, pytest is not intended to be run on its own, but our whole workflow uses tox, which is thus the only requirement in requirements-dev.txt. I see how this documentation line can be misunderstood, will clarify that pytest is needed to locally run tests.

Regarding R dependencies: Yes, each test set has some requirements, which can e.g. be found in the tox.ini file. I will add a clarification for that.

Indeed R needs a separate installation, as does julia. I will clarify that.
