### 2022-05-03 #4304 @blakeaw
Hi @osorensen, I have completed my review.

Overall, I think the submission looks good and can proceed with minor revisions. I raised one issue regarding the automated tests and I had one minor comment about the Software paper:

- https://github.com/ICB-DCM/pyABC/issues/565
- https://github.com/ICB-DCM/pyABC/issues/566
