### 2022-05-04 6f7c162
Improve test installation instructions; fixes #565
