+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@
 
 repos:
 - repo: https://github.com/psf/black
-  rev: 22.1.0
+  rev: 22.3.0
   hooks:
   - id: black
     description: The uncompromising code formatter
@@ -34,7 +34,7 @@ repos:
       args: [--py36-plus]
     - id: nbqa-isort
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.2.0
+  rev: v4.2.0
   hooks:
   - id: check-yaml
     description: Check yaml files for parseable syntax
