+++ tox.ini
@@ -157,7 +157,7 @@ deps =
     flake8-colors >= 0.1.6
     #flake8-commas >= 2.0.0
     flake8-comprehensions >= 3.2.3
-    flake8-print >= 3.1.4
+    flake8-print >= 5.0.0
     flake8-black >= 0.2.3
     flake8-isort >= 4.0.0
     # flake8-docstrings >= 1.5.0
