+++ tox.ini
@@ -150,7 +150,7 @@ skip_install = true
 deps =
     # TODO unfix when bandit dep fixed
     bandit < 1.7.3
-    black >= 22.1.0
+    black >= 22.3.0
     flake8 >= 3.8.3
     flake8-bandit >= 2.1.2
     flake8-bugbear >= 20.1.4
