### 2022-05-04 a84f9cf
Improve test installation instructions (#565); fix for flake8-print 5.0, black 22.3 (#567)
* Improve test installation instructions; fixes #565

* Fix for flake8-print 5.0, black 22.3

* update black version in tox
