### 2022-05-05 ICB-DCM/pyABC#570 @hpesonen
JOSS Review - comment on the Software paper
Hi!

Here's a few minor comments about the paper part of the submission

- I agree with Issue #566

-  L17-18 Clarify what type of noise is referred to in the following sentence “In particular, we implement algorithms to account for noise…”

- L34-35: Following sentence is incomplete  “In a nutshell, ABC circumvents likelihood evaluation by accepting if a distance between simulated and observed data is below a threshold”.

- L34: I’m not sure if if completely correct to describe the forward process model as a “black-box” model? Wouldn’t this make it impossible to model the parameters of the simulator a priori (as we do in ABC)?

- L40: I feel that the following sentence would benefit for declaring _why_ ABC is computationally expensive “While conceptually simple and widely available, ABC is computationally expensive…”

- Figure 1 : What does the illustration of a graph above the summary statistics in Figure 1 refer to? I’m also not sure what the illustration above “model” is either.

References:

- Some of the titles are incorrectly all lower case, e.g.  "Population growth of human y chromosomes…"

- In Lintusaari et al. names Kangasraasio and Jarvenpaa should be Kangasrääsiö and Järvenpää, respectively.

Additional comment:
---------------------

Plotting in the colab-notebooks is not currently working.
