### 2022-06-07 98b89ce
Update informative.ipynb (#575)
* Update informative.ipynb

fix typos on informative example

* Update conf.py

solving issue #576
