### 2022-06-20 4a3f5e7
address review comment "state of field" #566
