### 2022-06-20 #4304 @yannikschaelte
Dear all, I have now addressed the raised issues https://github.com/ICB-DCM/pyABC/issues/565 and https://github.com/ICB-DCM/pyABC/issues/566 (the latter so far only as a pull request in https://github.com/ICB-DCM/pyABC/pull/577).
