### 2022-06-21 #4304 @hpesonen
Hi @osorensen and @yannikschaelte! I've updated my checklist and I've checked the State of the field box based on https://github.com/ICB-DCM/pyABC/pull/577

I also had few minor comments on  https://github.com/ICB-DCM/pyABC/issues/570 that I hope would be addressed before checking the final box.
