+++ doc/joss_0_12/paper.bib
@@ -69,7 +69,7 @@
 
 @Article{AndrieuRob2009,
   author    = {Andrieu, Christophe and Roberts, Gareth O},
-  title     = {The pseudo-marginal approach for efficient Monte Carlo computations},
+  title     = {The pseudo-marginal approach for efficient {Monte} {Carlo} computations},
   number    = {2},
   pages     = {697--725},
   volume    = {37},
@@ -107,7 +107,7 @@
 
 @Article{PritchardSei1999,
   author    = {Pritchard, Jonathan K and Seielstad, Mark T and Perez-Lezaun, Anna and Feldman, Marcus W},
-  title     = {Population growth of human Y chromosomes: a study of Y chromosome microsatellites.},
+  title     = {Population growth of human {Y} chromosomes: a study of {Y} chromosome microsatellites.},
   number    = {12},
   pages     = {1791--1798},
   volume    = {16},
@@ -133,7 +133,7 @@
 @book{SissonFan2018Handbook,
 	author = {Sisson, Scott A and Fan, Yanan and Beaumont, Mark},
 	publisher = {Chapman and Hall/CRC},
-	title = {Handbook of approximate Bayesian computation},
+	title = {Handbook of approximate {B}ayesian computation},
 	year = {2018},
 }
 
@@ -167,7 +167,7 @@
 @InProceedings{DuttaSch2017,
   author        = {Dutta, Ritabrata and Schoengens, Marcel and Onnela, Jukka-Pekka and Mira, Antonietta},
   booktitle     = {Proceedings of the Platform for Advanced Scientific Computing Conference},
-  title         = {ABCpy: A User-Friendly, Extensible, and Parallel Library for Approximate {B}ayesian Computation},
+  title         = {{ABCpy}: A User-Friendly, Extensible, and Parallel Library for Approximate {B}ayesian Computation},
   doi           = {10.1145/3093172.3093233},
   isbn          = {978-1-4503-5062-4},
   location      = {Lugano, Switzerland},
@@ -183,8 +183,8 @@
 }
 
 @article{LintusaariVuo2018,
-  title   = {Elfi: Engine for likelihood-free inference},
-  author  = {Lintusaari, Jarno and Vuollekoski, Henri and Kangasraasio, Antti and Skyt{\'e}n, Kusti and Jarvenpaa, Marko and Marttinen, Pekka and Gutmann, Michael U and Vehtari, Aki and Corander, Jukka and Kaski, Samuel},
+  title   = {{Elfi}: Engine for likelihood-free inference},
+  author  = {Lintusaari, Jarno and Vuollekoski, Henri and Kangasr{\"a}{\"a}sio, Antti and Skyt{\'e}n, Kusti and Jarvenp{\"a}{\"a}, Marko and Marttinen, Pekka and Gutmann, Michael U and Vehtari, Aki and Corander, Jukka and Kaski, Samuel},
   journal = {Journal of Machine Learning Research},
   volume  = {19},
   number  = {16},
@@ -194,7 +194,7 @@
 
 @Article{TankhilevichIsh2020,
   author    = {Tankhilevich, Evgeny and Ish-Horowicz, Jonathan and Hameed, Tara and Roesch, Elisabeth and Kleijn, Istvan and Stumpf, Michael PH and He, Fei},
-  title     = {GpABC: a {J}ulia package for approximate {B}ayesian computation with {G}aussian process emulation},
+  title     = {{GpABC}: a {J}ulia package for approximate {B}ayesian computation with {G}aussian process emulation},
   number    = {10},
   pages     = {3286--3287},
   volume    = {36},
@@ -272,7 +272,7 @@
 
 @Article{KerrStu2021,
[…]
