### 2022-06-21 #4304 @hpesonen
Hi @yannikschaelte and @osorensen - I've reviewed the latest changes and I'm happy to check all the boxes on the checklist!

p.s. The comment above links to the issue 578 on JOSS-reviews
