### 2022-06-21 #4304 @yannikschaelte
> p.s. The comment above links to the issue 578 on JOSS-reviews

Ah, of course. Changed it!
