### 2022-06-21 ICB-DCM/pyABC#566 @blakeaw
Thanks, @yannikschaelte. Looks good to me.
