### 2022-06-21 #4304 @yannikschaelte
Hi @osorensen , I should have addressed all points:

* Tagged version: https://github.com/ICB-DCM/pyABC/releases/tag/0.12.5
* ZENODO archive: https://doi.org/10.5281/zenodo.6677826
