+++ doc/joss_0_12/paper.bib
@@ -131,14 +131,14 @@
 }
 
 @book{SissonFan2018Handbook,
-	author = {Sisson, Scott A and Fan, Yanan and Beaumont, Mark},
+	author = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
 	publisher = {Chapman and Hall/CRC},
 	title = {Handbook of approximate {B}ayesian computation},
 	year = {2018},
 }
 
 @Article{SissonFan2007,
-  author        = {Sisson, S. A. and Fan, Y. and Tanaka, M. M.},
+  author        = {Sisson, Scott A. and Fan, Y. and Tanaka, M. M.},
   title         = {Sequential {M}onte {C}arlo without likelihoods},
   number        = {6},
   pages         = {1760--1765},
@@ -412,7 +412,7 @@
 }
 
 @article{BlumNun2013,
-	author = {Blum, Michael GB and Nunes, Maria Antonieta and Prangle, Dennis and Sisson, Scott A},
+	author = {Blum, Michael GB and Nunes, Maria Antonieta and Prangle, Dennis and Sisson, Scott A.},
 	journal = {Statistical Science},
 	number = {2},
 	pages = {189--208},
