### 2022-06-25 #4304 @Kevin-Mattheus-Moerman
@yannikschaelte
- I checked the archive meta-data (title and authors) and all looks good. :+1:
- I checked the version tag for the review and archive and have the following comments:

- [ ] This review says it is currently for `0.12.4` however the [archive](https://zenodo.org/record/6677827) says `version 1`. If I inspect the project repository I see a tag for `0.12.5`. Can you please comment on what the version should be? The version should be the same for this review and the archive. I get the impression both should be `0.12.5` but do let me know which it should be. You may need to manually edit the ZENODO archive to have the correct version label.

- I proofread the paper and all seems fine :+1:
