### 2022-06-25 #4304 @yannikschaelte
Hi @Kevin-Mattheus-Moerman , thanks for proofreading! You are right, it should be 0.12.5 (I created a new version with the latest changes esp. from the review process). I updated the zenodo accordingly.
