### 2022-06-25 #4304 @Kevin-Mattheus-Moerman
@yannikschaelte congratulations on your publication in JOSS!

Thanks for editing this one @osorensen! And thank you @mattpitkin, @hpesonen, @blakeaw for your review efforts!!! :tada:
